Updated values, which i rounded for...some reason

+++ src/main/java/online/maestoso/soulforged/material/ArmorMaterial.java
@@ -1,7 +1,7 @@
 package online.maestoso.soulforged.material;
 
 public class ArmorMaterial extends Material {
-    public ArmorMaterial(int hardness, int workability, int density, int durability, int heat, int padding) {
+    public ArmorMaterial(double hardness, int workability, int density, int durability, int heat, int padding) {
         super(hardness, 0, workability, density, durability, heat, padding);
     }
