+++ src/main/java/online/maestoso/soulforged/material/CrystalMaterial.java
@@ -1,7 +1,7 @@
 package online.maestoso.soulforged.material;
 
 public class CrystalMaterial extends ToolMaterial {
-    public CrystalMaterial(int hardness, double edgeholding, int workability, int density, int durability) {
+    public CrystalMaterial(double hardness, double edgeholding, int workability, int density, int durability) {
         super(hardness, edgeholding, workability, density, durability, 0);
     }
 }
