+++ src/main/java/online/maestoso/soulforged/material/ToolMaterial.java
@@ -1,7 +1,7 @@
 package online.maestoso.soulforged.material;
 
 public class ToolMaterial extends Material {
-    public ToolMaterial(int hardness, double edgeholding, int workability, int density, int durability, int heat) {
+    public ToolMaterial(double hardness, double edgeholding, int workability, int density, int durability, int heat) {
         super(hardness, edgeholding, workability, density, durability, heat, 0);
     }
